+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.0.6
 - Fix Launch being stopped when using Biolith 1.0.1-beta.1
+- Pull Fabric Loot Modifier from LootTable Json
 
 # 1.0.5
 - Fix WailingTrigger Based Advancements triggering all at once